feat(Navigation): add CurrentPathActions [YTFRONT-4037]
