+++ packages/ui/src/ui/pages/navigation/content/File/File.scss
@@ -13,8 +13,4 @@
     &__actions {
         margin-bottom: 20px;
     }
-
-    &__download {
-        margin-right: 20px;
-    }
 }
