--- packages/ui/src/ui/pages/navigation/content/Table/TableOverview/EditTableAction.scss
@@ -1,5 +1,0 @@
-.edit-table-actions {
-    &__button {
-        margin-left: 10px;
-    }
-}
