+++ packages/ui/src/ui/pages/navigation/content/Table/TableOverview/TableOverview.scss
@@ -133,4 +133,8 @@ $overview-height: 50px;
             margin-left: auto;
         }
     }
+
+    &__path-actions {
+        margin-left: 10px;
+    }
 }
